# specs/openapi.yaml
openapi: 3.1.0
info:
  title: NarrativeOS API
  version: 0.3.0
paths:
...
  /v1/worlds:
    get:
      summary: List published worlds for the shelf
      responses:
        "200":
          description: Shelf worlds
    post:
      summary: Create or upsert a world bundle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [world_bible, event_atoms]
              properties:
                world_bible:
                  $ref: "./world_bible.schema.json"
                event_atoms:
                  type: array
                  items:
                    $ref: "./event_atom.schema.json"
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: World created
  /v1/sessions:
    get:
      summary: List sessions
      responses:
        "200":
          description: Session list
    post:
      summary: Create a new narrative session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [world_id, initial_state]
              properties:
                world_id:
                  type: string
                initial_state:
                  $ref: "./narrative_state.schema.json"
                player_profile:
                  type: object
                  additionalProperties: true
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Session created
  /v1/sessions/{session_id}/step:
    post:
      summary: Advance a session by one event-selection step
      parameters:
        - in: path
          name: session_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [player_input]
              properties:
                player_input:
                  type: string
                intent_override:
                  type: object
                  additionalProperties:
                    type: number
                candidate_events:
                  type: array
                  items:
                    $ref: "./event_atom.schema.json"
                beam_width:
                  type: integer
                depth:
                  type: integer
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Step result with candidates, critic trace, and rendered prose
  /v1/sessions/{session_id}/replay:
    get:
      summary: Replay a session with event, critic, and state traces
      parameters:
        - in: path
          name: session_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Replay payload
...
  /v1/routes/preview:
    post:
      summary: Preview best routes for a state and candidate event set
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [world, state, candidate_events]
              properties:
                world:
                  $ref: "./world_bible.schema.json"
                state:
                  $ref: "./narrative_state.schema.json"
                candidate_events:
                  type: array
                  items:
                    $ref: "./event_atom.schema.json"
                beam_width:
                  type: integer
                depth:
                  type: integer
      responses:
        "200":
          description: Route preview result with scored candidates and critic trace